code/requirements.txt
pycryptodome==3.10.1
CherryPy==18.6.0
pysaml2==6.5.1
bcrypt==3.2.0
ws4py==0.5.1
Jinja2==3.0.1
opencv-python==4.5.3.56